Split off a legacy mcrypt class
