+++ src/class-wordpress-options-panels.php
@@ -14,6 +14,7 @@ namespace WPOP\V_5_0;
  * Class WordPress_Options_Panels
  */
 class WordPress_Options_Panels {
+
 	/**
 	 * Load files required to use this utility.
 	 */
@@ -41,4 +42,5 @@ class WordPress_Options_Panels {
 			require_once $file;
 		}
 	}
+
 }
