+++ src/inc/class-panel.php
@@ -121,6 +121,12 @@ class Panel {
 			define( 'WPOP_ENCRYPTION_KEY', Password::pad_key( sha1( $trimmed_key, true ) ) );
 		}
 
+		if ( ! defined( 'WPOP_OPENSSL_ENCRYPTION_KEY' ) ) {
+			// IMPORTANT: If you don't define a key, the class hashes the AUTH_KEY found in wp-config.php,
+			// locking the encrypted value to the current environment.
+			define( 'WPOP_OPENSSL_ENCRYPTION_KEY', hash( 'sha256', wp_salt(), true ) );
+		}
+
 		// Establish panel id.
 		$this->id = preg_replace( '/_/', '-', $args['id'] );
